chore: Add vim syntax highlighting
The syntax file is written in the file "nuru.vim".
It contains basic syntax highlighting for "nuru".
`regex` is used so there is a chance that it will not be perfect.

The following are highlighted:

- Keywords
- Types (most)
- Operators (most)
- Comments (better but not yet there)

Only the definition is done, the actual highlighting (colours) is done
by vim itself.
A good theme is also a good investment in this case.

This syntax highlighting `MAY` not work with neovim.
No tests (opening and closing the files) were done.
I currently don't have neovim so the burden falls to the future entity.

Vim needs to be told of the syntax highlighting to actually work.
This can be done through:

```vim
au BufRead,BufBewFile *.nr set filetype=nuru
```

When the above line is inserted everything should work as
expected.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>