chore: Successfully seperate the builtins
The builtins are now in their own package.
This is to ensure extendable functions as opposed
to the previous method which they were in a "fixed" array.

Accessing the functions still uses an array as before.
This may be fixed to use a better way of accesing the
builtins but that will be done later.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>