fix: Add support for types
Now types are checked before appending to the 'new string'.
The '%' is truncated before being printed even if the syntax
turns out to be wrong.

An array is used to keep the type and subsequently checked against.
This is ineffective speed and perfomance wise.
A better solution _may be found_ in the future though.

>> No error handling yet

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>