chore: Add more format specifiers and docs
Now the code can handle gracefully:

- boolean
- arrays

Error handling has been enhanced to acceptable level (whatever that is).
Type mismatch is now reported appropriately.
If there is a mismatch in format specifiers to the actual arguments, then an
error is raised.
This only happens if the format specifiers are more than the passed arguments.
This doesn't apply the other way round.
This is consistent with how C handles this same situation in lax situation
(not tested in strict and golang though).

Changed the allowed specifiers to match their swahili names, making it easier
for swahili learners to pick up the specifiers.
This can be reverted though as it is mostly convention and not hard rules.

Code comments have been included to try and break down the code logic.
Repetetive part(s) have only been explained once.
This is to help maintainers and developers skim through the logic with the
help of the comments and understand it.
I know that this doensn't substitute clean readable code but it sure does help.

Format specifiers:

- %n -> Nambari
- %t -> Tungo (ama neno)
- %b -> Buliani (kweli, sikweli)
- %s -> Safu (ama orodha)

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>