chore: Add initial implementation of flags
Now flags infrastructure has been set up fully.
Addition of flags can be done now in few simple modifications.
For now only the following flags have been implemented:

- '+': Add a preceding '+' in integer values.
- '.': Add a predefined number of zeroes to float (DESIMALI)
- '!': Convert to uppercase the string passed (should be removed in the future).

The above implementated flags are just for demonstration of the flags
feature and not a concreate impelementation.

'%d' Have been added to the list of supported flags.
The flags represents "DESIMALI" type (e.g 2.4)

Error reporting is still not at par but good enough to nudge the user
in the right direction, but %.12 for example doesn't report any error.

The code still doesn't actually follow the DRY principle.
A lot of code is repeated.
Not sure if this is the biggest concern as of now.

Example usage as of now:

>>> andika('Habari %!t\n', 'nuru')
Habari NURU

>>> andika('%+.5d\n', 2.3)
+2.30000

If the flag is not known they it is silently ignored (the same as in golang (partly though)).
If the syntax is messed up, then undefined there will be undefied behaviour:
e.g:

>>> andika('%#.2_3d\n', 2.4)
2.40

This may or may not be the desired behaviour but the best that can be done (as of now).

NOTE:
Now andika() doesn't add an extra '\n' to the end of strings.
This **may** break the previous implementations.
This only happens if the first argument is a string, else the previous
implementation still aplies.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>