chore: Convert parser errors to 'Syntax Error'
The parser doesn't emit its own errors, it relies on 'errd' package to do so.
This makes it easier to format the error messages in a uniform parsable form.

Instead of:
	Mstari 1: Tumeshindwa kuparse HARAMU
Its:
	Kosa la kisintaksia: h.nr:1:4
	Tumeshindwa kuparse '#'

This gives the user the hint where the problem is and what caused the problem.

Parser errors doesn't exit after display.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>