chore: Add context line
The line in which the error occured in is now displayed along with the error.
This makes it easier to see what went wrong at a glance without opening the file to confirm.
A caret is placed at the start of the error token to help the user know where they can start fixing from.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>