fix: Make most of the tests pass
From previous changes, the tests started to fail because of added
<NEWLINE> token.

This is now fixed for most modules except the evaluator module.
The remaining problem may be related to the NEWLINE token.
The tests check the last token, which previously would be the value.
But now the last token would be the NEWLINE instead of another token
like INTEGER.

When running, the code works correctly.
This strongly indicates the problem to be in the tests instead of the
code (unlike the other tests).

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>