chore: Resolve `hapa` in various locations
Modules that are referenced by `hapa` e.g:
    tumia hapa::faili
can be read succesfully.

Add a new information to the token `Filename`.
This is used for resolving, so its empty in most parts.
its initialized by the import parser.

The stdlib will be referenced by `msingi` instead of `mkt`.
This is more approriate and sensible than the last.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>