fix: Allow all types to be exportable
Only functions were allowed to be imported/exported.
If a variable or a function name (without a call), then it will fail.
This restricted the way that packages can be written.
It meant that to export anything, you need to wrap it inside a function.

If a function that is not executed is printed, it follows the nodejs
format.
[Function: <name>] i.e [undo: <jina>].
If the function has parameters, then they can are displayed also.
[undo: <jina>; inahitaji: <[vinavyohitajika]>]
This is more cleaner than trying to piece together the function body.

Add extra logic for when assignment is to a function.
This is to add the function name which was lost when handing over from
assign to functionEval.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>