chore: Add install script
The install script allows user to install the nuru executabe in
'correct' location. [1]
Installation only uses the `nuru` executabe and discards the rest of the
artifacts.

The 'other' artifacts are `LICENSE` and `README.md` files.
They are not useful as they contain little value to the end user.
Proper documentation should be provided [1].

Issues:
    [1] Improve Installation Structure & Uninstallation Process
	- https://github.com/NuruProgramming/Nuru/issues/100

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>